make in handlers.go:  detete, post, user register change id in DB from integer to bigserial

+++ migrations/20220712200247_UserCreationMigrations.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id integer not null primary key,
+    id bigserial not null primary key,
     login varchar not null unique,
     password varchar not null
 );
